Use correct epoch context to get signature set
