+++ packages/lodestar-beacon-state-transition/src/fast/block/processAttesterSlashing.ts
@@ -1,4 +1,4 @@
-import {AttesterSlashing, BeaconState} from "@chainsafe/lodestar-types";
+import {AttesterSlashing, BeaconState, ValidatorIndex} from "@chainsafe/lodestar-types";
 
 import {isSlashableValidator, isSlashableAttestationData} from "../../util";
 import {EpochContext} from "../util";
@@ -27,7 +27,7 @@ export function processAttesterSlashing(
   let slashedAny = false;
   const attSet1 = new Set(attestation1.attestingIndices);
   const attSet2 = new Set(attestation2.attestingIndices);
-  const indices = [];
+  const indices: ValidatorIndex[] = [];
   for (const i of attSet1.values()) {
     if (attSet2.has(i)) {
       indices.push(i);
